+++ packages/elements/src/handlers/trap/types.ts
@@ -1,0 +1,5 @@
+export type TrapState = {
+	element: HTMLElement;
+	previousActiveElement: HTMLElement | null;
+	siblingElements: HTMLElement[];
+};