+++ packages/elements/CHANGELOG.md
@@ -9,6 +9,6 @@
 - Event handler now supports registering event handlers on `document`, `body` and `head` by prefixing the event name with the target. Ie. `data-handler--event.document.click="myAction"`.
 - Breaking change: all state values in `data-bind--` attributes must be prefixed with `$`. Ie. `data-bind--aria-label="scope:$state"`.
 - All first-party handlers now support using state as well as actions. Just prefix the value with either a `$` or `@` to denote the member type.
-
+- Added a new first party handler for trapping focus. This can be used via the `data-handler--trap` attribute and supports both member types.